Created RVD.
wip.  Compiling, untested.
