+++ src/main/scala/is/hail/methods/LinearRegression.scala
@@ -4,7 +4,6 @@ import breeze.linalg._
 import breeze.numerics.sqrt
 import is.hail.annotations._
 import is.hail.expr._
-import is.hail.sparkextras.WritableRegionValue
 import is.hail.stats._
 import is.hail.utils._
 import is.hail.variant._
